+++ CHANGELOG.md
@@ -22,3 +22,6 @@
 
 **March 23rd, 2022**
  - (DCh) in-game menu fix, minor code refactoring
+
+*March 25th, 2022*
+- (KTs) added starting screen, dialog window.